requirements.txt
pyrogram
python-dotenv
TgCrypto

